removed edit trans modal content and renamed AddTransactionsModalContent to TransactionsModalContent
