--- components/modal/modal_content/UpdateTransactionModalContent.tsx
@@ -1,50 +1,0 @@
-import { isValidAmount, isValidDate, isValidDescription } from "@/app/(tabs)/utilities/balance-utilities";
-import { TransactionModalContentProps } from "@/app/types";
-import { AppThemedText } from "@/components/app_components/AppThemedText";
-import AppThemedTextInput from "@/components/app_components/AppThemedTextInput";
-
-
-const EditTransactionModalContent = ({
-    amount,
-    date,
-    description,
-    setAmount,
-    setDate,
-    setDescription,
-    mutation,
-    setModalVisible,
-  }: TransactionModalContentProps) => {
-    return (
-      <>
-        <AppThemedTextInput
-          checkValue={isValidDate}
-          iconName="calendar"
-          placeholder="MM/DD/YY"
-          secureEntry={false}
-          setValue={setDate}
-          value={date}
-        />
-        <AppThemedTextInput
-          checkValue={isValidAmount}
-          placeholder="Amount"
-          secureEntry={false}
-          setValue={setAmount}
-          value={amount}
-        />
-        <AppThemedTextInput
-          checkValue={isValidDescription}
-          placeholder="Description"
-          secureEntry={false}
-          setValue={setDescription}
-          value={description}
-        />
-        <AppThemedText type="link" onPress={() => mutation.mutate()}>
-          Add
-        </AppThemedText>
-        <AppThemedText type="link" onPress={() => setModalVisible(false)}>
-          Close
-        </AppThemedText>
-      </>
-    );
-  };
-export default EditTransactionModalContent;
